Changed packaging for gui classes
