+++ desktop-tool/src/main/java/ch/hsr/zebrastreifensafari/view/screen/MainGUI.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="ch.hsr.zebrastreifensafari.gui.MainGUI">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="ch.hsr.zebrastreifensafari.view.screen.MainGUI">
   <grid id="27dc6" binding="mainPanel" layout-manager="BorderLayout" hgap="0" vgap="0">
     <constraints>
       <xy x="20" y="20" width="751" height="603"/>
@@ -83,7 +83,7 @@
                   <size top="0" left="0" bottom="0" right="0"/>
                 </border>
                 <children>
-                  <component id="33ce1" class="ch.hsr.zebrastreifensafari.gui.CrossingTable" binding="crossingTable" custom-create="true">
+                  <component id="33ce1" class="ch.hsr.zebrastreifensafari.view.table.CrossingTable" binding="crossingTable" custom-create="true">
                     <constraints/>
                     <properties/>
                   </component>
@@ -105,7 +105,7 @@
                 <properties/>
                 <border type="none"/>
                 <children>
-                  <component id="65af1" class="ch.hsr.zebrastreifensafari.gui.RatingTable" binding="ratingTable" custom-create="true">
+                  <component id="65af1" class="ch.hsr.zebrastreifensafari.view.table.RatingTable" binding="ratingTable" custom-create="true">
                     <constraints/>
                     <properties>
                       <rowSelectionAllowed value="true"/>
