+++ desktop-tool/src/main/java/ch/hsr/zebrastreifensafari/view/screen/MainGUI.java
@@ -1,15 +1,18 @@
-package ch.hsr.zebrastreifensafari.gui;
+package ch.hsr.zebrastreifensafari.view.screen;
 
 import ch.hsr.zebrastreifensafari.controller.AboutController;
-import ch.hsr.zebrastreifensafari.gui.create.CreateCrossingGUI;
-import ch.hsr.zebrastreifensafari.gui.create.CreateRatingGUI;
-import ch.hsr.zebrastreifensafari.gui.edit.EditCrossingGUI;
-import ch.hsr.zebrastreifensafari.gui.edit.EditRatingGUI;
 import ch.hsr.zebrastreifensafari.jpa.entities.*;
 import ch.hsr.zebrastreifensafari.model.Model;
 import ch.hsr.zebrastreifensafari.service.DataServiceLoader;
 import ch.hsr.zebrastreifensafari.service.Properties;
 import ch.hsr.zebrastreifensafari.service.WebsiteService;
+import ch.hsr.zebrastreifensafari.view.component.JTextPlaceHolder;
+import ch.hsr.zebrastreifensafari.view.screen.modify.create.CreateCrossingGUI;
+import ch.hsr.zebrastreifensafari.view.screen.modify.create.CreateRatingGUI;
+import ch.hsr.zebrastreifensafari.view.screen.modify.edit.EditCrossingGUI;
+import ch.hsr.zebrastreifensafari.view.screen.modify.edit.EditRatingGUI;
+import ch.hsr.zebrastreifensafari.view.table.CrossingTable;
+import ch.hsr.zebrastreifensafari.view.table.RatingTable;
 import org.eclipse.persistence.exceptions.DatabaseException;
 
 import javax.persistence.EntityNotFoundException;
