+++ desktop-tool/src/main/java/ch/hsr/zebrastreifensafari/view/screen/modify/ModifyGUI.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="ch.hsr.zebrastreifensafari.gui.ModifyGUI">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="ch.hsr.zebrastreifensafari.view.screen.modify.ModifyGUI">
   <grid id="27dc6" binding="mainPanel" layout-manager="BorderLayout" hgap="0" vgap="0">
     <constraints>
       <xy x="20" y="20" width="683" height="484"/>
