+++ desktop-tool/src/main/java/ch/hsr/zebrastreifensafari/view/screen/modify/create/CreateCrossingGUI.java
@@ -1,7 +1,7 @@
-package ch.hsr.zebrastreifensafari.gui.create;
+package ch.hsr.zebrastreifensafari.view.screen.modify.create;
 
-import ch.hsr.zebrastreifensafari.gui.MainGUI;
-import ch.hsr.zebrastreifensafari.gui.ModifyGUI;
+import ch.hsr.zebrastreifensafari.view.screen.MainGUI;
+import ch.hsr.zebrastreifensafari.view.screen.modify.ModifyGUI;
 import ch.hsr.zebrastreifensafari.jpa.entities.Crossing;
 import ch.hsr.zebrastreifensafari.jpa.entities.Rating;
 import ch.hsr.zebrastreifensafari.service.DataServiceLoader;
